추가요청API(오늘의 도전 미믹 +도전자) + 방문자수 카운트 기능 추가된것 확인했습니다.

+++ controllers/socialLogin.controller.js
@@ -1,6 +1,8 @@
 const { User } = require("../models");
 const jwt = require("jsonwebtoken");
 const passport = require("passport");
+const Joi = require("../advice/joi");
+const Boom = require("@hapi/boom");
 require("dotenv").config();
 
 // Kakao callback Controller
@@ -23,11 +25,20 @@ exports.kakaoLogin = (req, res, next) => {
   )(req, res, next);
 };
 
+// passport-kakao 연결끊기 콜백 API
 exports.deleteKakao = async (req, res, next) => {
+  const joi = new Joi();
   try {
-    const { user_id } = req.query;
+    const { user_id } = await joi.idSchema.validateAsync(req.query);
+    const user = await User.findOne({ where: { snsId: user_id } });
+    if (!user) {
+      throw Boom.badRequest("존재하지 않는 카카오 로그인 회원입니다");
+    }
 
     await User.destroy({ where: { snsId: user_id } });
+    res.status(201).json({
+      message: "success(kakao)",
+    });
   } catch (err) {
     next(err);
   }
